Removed some unused methods from client (now in handlesystemconnector).
