+++ b2handle/tests/handleconnector_writeaccess_patched_test.py
@@ -10,8 +10,7 @@ import mock
 from mock import patch
 sys.path.append("../..")
 import b2handle.handlesystemconnector as connector
-from b2handle.handleexceptions import HandleSyntaxError
-from b2handle.handleexceptions import CredentialsFormatError
+from b2handle.handleexceptions import HandleSyntaxError, CredentialsFormatError, GenericHandleError, HandleNotFoundException
 from b2handle.util import check_handle_syntax, check_handle_syntax_with_index, remove_index_from_handle
 from mockresponses import MockResponse, MockSearchResponse
 from utilities import replace_timestamps, failure_message
@@ -177,3 +176,53 @@ class EUDATHandleConnectorWriteaccessPatchedTestCase(unittest.TestCase):
         self.assertIn('Basic ', headers['Authorization'],
             'Authorization header not sent correctly: '+headers['Authorization'])
 
+    # check if username exists
+
+    @patch('requests.Session.get')
+    def test_check_if_username_exists_normal(self, getpatch):
+        """Test whether username exists."""
+
+        # Test variables
+        handlerecord_string = open('resources/handlerecord_for_reading.json').read()
+        handlerecord_json = json.loads(handlerecord_string)
+        testhandle = '100:'+handlerecord_json['handle']
+
+        # Define the replacement for the patched method:
+        mock_response = MockResponse(success=True, content=handlerecord_string)
+        getpatch.return_value = mock_response
+
+        # Call method and check result:
+        res = self.inst.check_if_username_exists(testhandle)
+        self.assertTrue(res,
+            'The handle exists, so "check_if_username_exists" should return true!')
+
+    @patch('requests.Session.get')
+    def test_check_if_username_exists_inconsistent_info(self, getpatch):
+        """Test exception when contradictory inputs are given."""
+    
+        # Test variables
+        handlerecord_string = open('resources/handlerecord_for_reading.json').read()
+        testhandle = 'who/cares'
+
+        # Define the replacement for the patched method:
+        mock_response = MockResponse(success=True, content=handlerecord_string)
+        getpatch.return_value = mock_response
+
+        # Call method and check result:
+        with self.assertRaises(GenericHandleError):
+            self.inst.check_if_username_exists(testhandle)
+
+    @patch('requests.Session.get')
+    def test_check_if_username_exists_it_doesnot(self, getpatch):
+        """Test exception"""
+
+        # Test variables
+        testhandle = 'who/cares'
+        
+        # Define the replacement for the patched method:
+        mock_response = MockResponse(notfound=True)
+        getpatch.return_value = mock_response
+
+        # Call method and check result:
+        with self.assertRaises(HandleNotFoundException):
+            self.inst.check_if_username_exists(testhandle)
